requirements.txt
streamlit==1.30.0
PyPDF2==3.1.0
transformers==4.34.0
torch==2.1.0
